requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.2.0
audioread==3.0.1
bitarray==2.6.0
Brotli==1.1.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
coverage==7.4.3
datasets==2.18.0
decorator==5.1.1
dill==0.3.8
Django==5.0.3
djangorestframework==3.14.0
docx2txt==0.8
easyocr==1.7.1
exceptiongroup==1.2.0
fairseq2==0.2.0
fairseq2n==0.2.0
fasttext==0.9.2
fasttext-langdetect==1.0.5
filelock==3.9.0
fire==0.6.0
flake8==7.0.0
frozenlist==1.4.1
fsspec==2024.2.0
huggingface-hub==0.21.4
idna==3.6
imageio==2.34.0
iniconfig==2.0.0
Jinja2==3.1.2
jiwer==3.0.3
joblib==1.3.2
langdetect==1.0.9
lazy_loader==0.3
librosa==0.10.1
lingua-language-detector==2.0.2
llvmlite==0.42.0
lxml==5.1.0
MarkupSafe==2.1.3
mccabe==0.7.0
more-itertools==10.2.0
mpmath==1.3.0
msgpack==1.0.8
multidict==6.0.5
multiprocess==0.70.16
mutagen==1.47.0
networkx==3.2.1
ninja==1.11.1.1
numba==0.59.0
numpy==1.26.3
nvidia-cublas-cu11==11.11.3.6
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu11==11.8.87
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu11==11.8.89
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu11==11.8.89
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu11==8.7.0.84
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu11==10.9.0.58
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu11==10.3.0.86
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu11==11.4.1.48
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu11==11.7.5.86
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu11==2.19.3
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.4.99
nvidia-nvtx-cu11==11.8.86
nvidia-nvtx-cu12==12.1.105
openai-whisper==20231117
opencv-python-headless==4.9.0.80
overrides==7.7.0
packaging==23.2
pandas==2.2.1
pdf2image==1.17.0
pillow==10.2.0
platformdirs==4.2.0
pluggy==1.4.0
pooch==1.8.1
portalocker==2.8.2
pyarrow==15.0.1
pyarrow-hotfix==0.6
pybind11==2.11.1
pyclipper==1.3.0.post5
pycodestyle==2.11.1
pycparser==2.21
pycryptodomex==3.20.0
pydub==0.25.1
pyflakes==3.2.0
PyPDF2==3.0.1
pytesseract==0.3.10
pytest==8.1.1
pytest-cov==4.1.0
pytest-flake8==1.1.1
python-bidi==0.4.2
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
rapidfuzz==3.6.2
regex==2023.12.25
requests==2.31.0
sacrebleu==2.4.0
safetensors==0.4.2
scikit-image==0.22.0
scikit-learn==1.4.1.post1
scipy==1.12.0
seamless-communication @ git+https://github.com/facebookresearch/seamless_communication.git@75ed7ef2db2bcc66deb794e19626127ca10cfb4a
sentencepiece==0.2.0
shapely==2.0.3
simuleval==1.1.4
six==1.16.0
sonar-space==0.2.1
soundfile==0.12.1
sox==1.4.1
soxr==0.3.7
sqlparse==0.4.4
sympy==1.12
tabulate==0.9.0
tbb==2021.11.0
termcolor==2.4.0
TextGrid==1.6.1
threadpoolctl==3.3.0
tifffile==2024.2.12
tiktoken==0.6.0
tokenizers==0.15.2
tomli==2.0.1
torch==2.1.2+cu118
torchaudio==2.1.2+cu118
torcheval==0.0.7
torchvision==0.16.2+cu118
tornado==6.4
tqdm==4.64.1
transformers==4.38.2
triton==2.1.0
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
websockets==12.0
xxhash==3.4.1
yarl==1.9.4
yt-dlp==2024.3.10
